% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ci_plots.R
\name{add_CI_plot}
\alias{add_CI_plot}
\title{Add confidence interval(s) to an existing fitted distribution ggplot}
\usage{
add_CI_plot(base_plot, bts, logscale, CI.level = 0.95)
}
\arguments{
\item{base_plot}{A ggplot object that depicts the fit(s) of the bootstrap(s).}

\item{bts}{A list of \code{bootdist} or \code{bootdistcens} objects.}

\item{logscale}{if \code{TRUE}, uses a logarithmic scale for the \eqn{x}-axis}

\item{CI.level}{A strictly positive numeric smaller than 1. The level of the
confidence interval(s).}
}
\value{
A ggplot object.
}
\description{
Add confidence interval(s) to an existing fitted distribution ggplot
}
