% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bootstraps.R
\name{bootdist_fun}
\alias{bootdist_fun}
\title{Determines the appropriate bootstrap function based on a fit object}
\usage{
bootdist_fun(x, ft, niter = 200)
}
\arguments{
\item{x}{Not used}

\item{ft}{A \code{fitdist} or \code{fitdistcens} object}

\item{niter}{An integer. The number of iterations set for the bootstrap
function}
}
\value{
An appropriate bootstrap function.
}
\description{
Determines the appropriate bootstrap function based on a fit object
}
