% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_code_r.R
\name{code_r_ssd}
\alias{code_r_ssd}
\title{Generate a script according to user in-app inputs}
\usage{
code_r_ssd(
  data,
  distributions,
  censored = FALSE,
  logscale = TRUE,
  unit = "arbitrary unit",
  names = FALSE,
  groups = FALSE,
  CI.level = 0.95
)
}
\arguments{
\item{data}{If \code{censored = TRUE}, a data.frame with columns \code{left}
and \code{right}. If \code{censored = FALSE}, a data.frame with a column
\code{conc}.}

\item{distributions}{A list of the names of the distributions to apply (e.g.,
\code{'lnorm'}, \code{'llogis'})}

\item{censored}{A Boolean, whether the given data is to be interpreted as
censored}

\item{logscale}{A logical}

\item{unit}{A character string}

\item{names}{A logical}

\item{groups}{A logical}

\item{CI.level}{A numerical}
}
\value{
A character string containing a R script.
}
\description{
Generate a script according to user in-app inputs
}
