% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bootstraps.R
\name{combine_boot_samples}
\alias{combine_boot_samples}
\title{Combine several bootstrap samples into a single \code{bootdist} or
\code{bootdistcens} object}
\usage{
combine_boot_samples(bs)
}
\arguments{
\item{bs}{A list of \code{bootdist} or \code{bootdistcens} objects
obtained from the same fit object.}
}
\value{
A \code{bootdist} or \code{bootdistcens} object with a number of
  samples equal to the sum of number of samples of each bootstrap from the
  list input.
}
\description{
Combine several bootstrap samples into a single \code{bootdist} or
\code{bootdistcens} object
}
