% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{custom_theme}
\alias{custom_theme}
\title{Custom ggplot theme created for MOSAIC bioacc overwriting \code{theme_bw}}
\usage{
custom_theme()
}
\description{
Custom ggplot theme created for MOSAIC bioacc overwriting \code{theme_bw}
}
