% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endosulfan}
\alias{endosulfan}
\title{Summary of 48 to 96-hour acute toxicity values for endosulfan}
\format{
\code{endosulfan}

A data frame with 88 rows and 3 columns:
\describe{
  \item{conc}{Lethal or effective concentration in \eqn{\mu}g/L.}
  \item{name}{Specie's name for each concentration.}
  \item{group}{Classification by geographical origin and type: fish or
  arthropod.}
}
}
\source{
\doi{10.1007/s00244-003-3212-5}
}
\usage{
endosulfan
}
\description{
Summary of 48 to 96-hour acute toxicity values (LC50 and EC50 values) for
exposure of Australian and Non-Australian taxa to endosulfan.
}
\keyword{datasets}
