% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fluazinam}
\alias{fluazinam}
\title{48-hour acute toxicity values for fluazinam}
\format{
\code{fluazinam}

A data frame with 14 rows and 4 columns:
\describe{
  \item{left, right}{Lower & higher bounds of the effective concentration
  interval in \eqn{\mu}g/L.}
  \item{name}{Specie's name for each effective concentration.}
  \item{group}{Classification of the specie, not all rows use the same
  taxonomic rank.}
}
}
\source{
\doi{10.1016/j.ecoenv.2009.09.019 }
}
\usage{
fluazinam
}
\description{
48-hour acute toxicity values (EC50 values) for exposure of
macroinvertebrates and zooplancton to fluazinam.
}
\keyword{datasets}
