% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bootstraps.R
\name{get_bootstrap}
\alias{get_bootstrap}
\title{Generate bootstrap sample(s) for a list of fit(s) and check their convergence}
\usage{
get_bootstrap(fits)
}
\arguments{
\item{fits}{A list of \code{fitdist} or \code{fitdistcens} objects}
}
\value{
A list of two lists. The first list contains the bootstrap sample(s)
and the second one the logical value of convergence.
}
\description{
Generate bootstrap sample(s) for a list of fit(s) and check their convergence
}
