% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_shiny_outputs.R
\name{get_custom_HCx}
\alias{get_custom_HCx}
\title{Get HCx values for a given x from a bootstrap}
\usage{
get_custom_HCx(x, distributions, bootstrap = NULL, CI.level = 0.95)
}
\arguments{
\item{x}{An integer between 0 and 100.The percent of the hazardous
concentration desired (HCx).}

\item{distributions}{A list of the names of the distributions to apply (e.g.,
\code{'lnorm'}, \code{'llogis'})}

\item{bootstrap}{A list of \code{bootdist} or \code{bootdistcens} objects
corresponding to the fits provided.}

\item{CI.level}{A numeric, either 0.95 or 0.9. The level of the
confidence interval(s).}
}
\value{
A string describing the HCx and its confidence interval for each
distribution.
}
\description{
Get HCx corresponding to the provided x for each fit provided with confidence
intervals
}
