% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fits.R
\name{get_fits}
\alias{get_fits}
\title{Fit the specified distributions to the given data}
\usage{
get_fits(data, distributions, censored)
}
\arguments{
\item{data}{If \code{censored = TRUE}, a data.frame with columns \code{left}
and \code{right}. If \code{censored = FALSE}, a data.frame with a column
\code{conc}.}

\item{distributions}{A list of the names of the distributions to apply (e.g.,
\code{'lnorm'}, \code{'llogis'})}

\item{censored}{A Boolean, whether the given data is to be interpreted as
censored}
}
\value{
A list containing the fit object(s) of class \code{fitdist} or
\code{fitdistcens}.
}
\description{
Fit the specified distributions to the given data
}
