% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{get_xlab}
\alias{get_xlab}
\title{Create a label for x axis based on concentration unit and log scale}
\usage{
get_xlab(unit, logscale)
}
\arguments{
\item{unit}{A character vector, the unit of the toxicity data}

\item{logscale}{if \code{TRUE}, uses a logarithmic scale for the \eqn{x}-axis}
}
\value{
A character vector.
}
\description{
Create a label for x axis based on concentration unit and log scale
}
