% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_grouped_plots.R
\name{group_cdf_uncensored}
\alias{group_cdf_uncensored}
\title{Graphical representation of grouped toxicity uncensored data}
\usage{
group_cdf_uncensored(data, p)
}
\arguments{
\item{data}{A data.frame with columns \code{conc} and \code{group}.}

\item{p}{a ggplot object showing the empirical cumulative distribution
function of the uncensored data provided.}
}
\value{
A ggplot object.
}
\description{
Add a group coloration to a provided empirical cdf representation according
to a "group" column in data.
}
