% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_llogis_functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dllogis}
\alias{pllogis}
\alias{qllogis}
\alias{rllogis}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{actuar}{\code{\link[actuar:Loglogistic]{dllogis}}, \code{\link[actuar:Loglogistic]{pllogis}}, \code{\link[actuar:Loglogistic]{qllogis}}, \code{\link[actuar:Loglogistic]{rllogis}}}
}}

