% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salinity_order}
\alias{salinity_order}
\title{72-hour acute salinity tolerance of macro-invertebrates grouped by order.}
\format{
\code{Salinity_order}

A data frame with 108 rows and 3 columns:
\describe{
   \item{left, right}{Lower & higher bounds of the lethal concentration
  interval in mS/cm\eqn{^2}.}
  \item{group}{Order of each tested specie.}
}
}
\source{
\doi{10.1139/f06-080}
}
\usage{
salinity_order
}
\description{
72-hour acute salinity tolerance (LC50 values) of riverine
macro-invertebrates grouped by taxonomic order.
}
\keyword{datasets}
