% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_custom_shiny_inputs.R
\name{switchInput}
\alias{switchInput}
\title{Custom toggle switch input for shiny UI}
\usage{
switchInput(id, aria_label, checked = TRUE)
}
\arguments{
\item{id}{The \code{input} slot that will be used to access the value.}

\item{aria_label}{An invisible label for screen readers.}

\item{checked}{Whether to create the switch as ON.}
}
\value{
A toggle switch control that can be added to a UI definition.
}
\description{
Custom toggle switch input for shiny UI
}
