% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_bootstraps.R
\name{test_conv}
\alias{test_conv}
\title{Test the approximate equality of the quantiles from several bootstrap samples}
\usage{
test_conv(bs, probs)
}
\arguments{
\item{bs}{A list of \code{bootdist} or \code{bootdistcens} objects
obtained from the same fit object.}

\item{probs}{A numeric vector of probabilities with values in [0,1]}
}
\value{
A logical.
}
\description{
Test the approximate equality of the quantiles from several bootstrap samples
}
